refactor(api): extract Scheduler API to web/src/api/scheduler.ts
After rebasing on main, web/src/api/client.ts hit 1522 LOC and tripped
the file-size cap (1500 LOC). Move the Scheduler/Routines wire shapes
and helpers to a sibling module and re-export them from client.ts for
back-compat. client.ts drops to 1327 LOC.
